backend/requirements.txt
flask
flask-cors
python-dotenv
google-search-results
openai